rename "enums" to "definitions" and write modules in snake case
